initialized store in App and pass it as props to children
